Here doc working.
Here doc works now by creating a temp file and using GNL to check stdin input and save to /tmp/pipex, then reading from temp file as if it was an input file. Tweaked parsing a little as well. and some of the error handling. Norm is good as well. Now just need extensive testing prior to push.