Fixed mandatory.
Mandatory portion was still like before I update the pid linked list. I made a copy of bonus and put it in mandatory. Now I need to remove here_doc support for mandatory. Then I think I'm ready for push. Still some issues with error codes, but I dont think it will be necessary for me to get all the return codes exactly right. If its an issue, Ill come back and fix it. But I'm comfortable with that for now.